requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
anyio==3.6.2
apache-airflow==2.5.2
apache-airflow-providers-amazon==7.3.0
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.2.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==3.0.4
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
beautifulsoup4==4.12.0
blinker==1.5
boto==2.49.0
boto3==1.26.96
botocore==1.29.96
branca==0.6.0
cachelib==0.9.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
clickclick==20.10.2
cligj==0.7.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
crcmod==1.7
cron-descriptor==1.2.35
croniter==1.3.8
cryptography==39.0.2
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
exceptiongroup==1.1.1
fasteners==0.18
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.10.0
Fiona==1.9.1
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
folium==0.14.0
frozenlist==1.3.3
future==0.18.3
gcs-oauth2-boto-plugin==3.0
geopandas==0.12.2
google-apitools==0.5.32
google-auth==2.16.3
google-reauth==0.1.1
graphviz==0.20.1
gsutil==5.21
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
idna==3.4
importlib-metadata==4.13.0
inflection==0.5.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.3
jsonschema==4.17.3
jupyter==1.0.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
llvmlite==0.39.1
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
mypy-boto3-appflow==1.26.78
mypy-boto3-rds==1.26.72
mypy-boto3-redshift-data==1.26.88
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauth2client==4.1.3
openai==0.27.2
openai-whisper==20230314
packaging==23.0
pandas==1.5.3
pathspec==0.9.0
pendulum==2.1.2
plotly==5.13.0
pluggy==1.0.0
ply==3.11
prison==0.2.1
psutil==5.9.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==23.1.0
pyparsing==3.0.9
pyproj==3.4.1
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.7.1
pytzdata==2020.1
pyu2f==0.1.5
PyYAML==6.0
redshift-connector==2.0.910
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
retry-decorator==1.1.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.2
rsa==4.7.2
s3transfer==0.6.0
scramp==1.4.4
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-redshift==0.8.12
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.1
termcolor==2.2.0
text-unidecode==1.3
tiktoken==0.3.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.14
watchtower==2.0.1
Werkzeug==2.2.3
whisper==1.1.10
wrapt==1.15.0
WTForms==3.0.1
yarl==1.8.2
zipp==3.15.0
